% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{auto_max_diff}
\alias{auto_max_diff}
\title{Auto-calculate Maximum Difference for Normalization}
\usage{
auto_max_diff(values)
}
\arguments{
\item{values}{Numeric vector or list of values}
}
\value{
Numeric value for maximum difference in normalization calculations
}
\description{
Auto-calculate Maximum Difference for Normalization
}
\keyword{internal}
