% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R, R/2-similar-base-methods.R
\name{average_similarity}
\alias{average_similarity}
\title{Calculate Average Similarity Scores}
\usage{
average_similarity(x, ...)

average_similarity(x, ...)
}
\arguments{
\item{x}{A similarity object}

\item{...}{Additional arguments (not used)}
}
\value{
A named numeric vector of mean similarity scores for each method

A named numeric vector of mean similarity scores for each method
}
\description{
Calculates and returns the average similarity score for each method used in the comparison.
}
