% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same-factor.R
\name{average_similarity.same_factor}
\alias{average_similarity.same_factor}
\title{Print Method for same_factor Objects}
\usage{
\method{average_similarity}{same_factor}(x, ...)
}
\arguments{
\item{x}{A same_factor object}

\item{...}{Additional arguments passed to specific methods}
}
\value{
Invisibly returns the input object

Calculate Average Similarity Scores for same_factor Objects

A named numeric vector of mean similarity scores for each method
}
\description{
Calculates and returns the average similarity score for each method used in the comparison.
}
