% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same-number.R
\name{average_similarity.same_number}
\alias{average_similarity.same_number}
\title{Print Method for summary.similar_number Objects}
\usage{
\method{average_similarity}{same_number}(x, ...)
}
\arguments{
\item{x}{A same_number object}

\item{...}{Additional arguments passed to specific methods}
}
\value{
Invisibly returns the input object

Calculate Average Similarity Scores for same_number Objects

A named numeric vector of mean similarity scores for each method
}
\description{
Calculates and returns the average similarity score for each method used in the comparison.
}
