% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-s3-bridge.R
\name{average_similarity.similar}
\alias{average_similarity.similar}
\title{Calculate average similarity scores}
\usage{
\method{average_similarity}{similar}(x, ...)
}
\arguments{
\item{x}{A similarity object}

\item{...}{Additional arguments (not used)}
}
\value{
A named numeric vector of mean similarity scores for each method
}
\description{
Calculate average similarity scores
}
