% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-base.R
\name{average_similarity.SimilarityBase}
\alias{average_similarity.SimilarityBase}
\title{Method implementation for average_similarity to work with SimilarityBase}
\usage{
\method{average_similarity}{SimilarityBase}(x, ...)
}
\description{
Method implementation for average_similarity to work with SimilarityBase
}
