% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-calculators.R
\name{FactorSimilarityCalculator}
\alias{FactorSimilarityCalculator}
\title{Factor Similarity Calculator}
\description{
Calculator for factor/categorical similarity metrics
}
\section{Super class}{
\code{\link[samesies:SimilarityCalculatorBase]{samesies::SimilarityCalculatorBase}} -> \code{FactorSimilarityCalculator}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{levels}}{Combined levels from all factor inputs}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FactorSimilarityCalculator-get_type}{\code{FactorSimilarityCalculator$get_type()}}
\item \href{#method-FactorSimilarityCalculator-get_valid_methods}{\code{FactorSimilarityCalculator$get_valid_methods()}}
\item \href{#method-FactorSimilarityCalculator-calculate_similarity}{\code{FactorSimilarityCalculator$calculate_similarity()}}
\item \href{#method-FactorSimilarityCalculator-calculate_scores}{\code{FactorSimilarityCalculator$calculate_scores()}}
\item \href{#method-FactorSimilarityCalculator-validate_inputs}{\code{FactorSimilarityCalculator$validate_inputs()}}
\item \href{#method-FactorSimilarityCalculator-compare}{\code{FactorSimilarityCalculator$compare()}}
\item \href{#method-FactorSimilarityCalculator-clone}{\code{FactorSimilarityCalculator$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="samesies" data-topic="SimilarityCalculatorBase" data-id="validate"><a href='../../samesies/html/SimilarityCalculatorBase.html#method-SimilarityCalculatorBase-validate'><code>samesies::SimilarityCalculatorBase$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorSimilarityCalculator-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-FactorSimilarityCalculator-get_type}{}}}
\subsection{Method \code{get_type()}}{
Get the type of calculator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorSimilarityCalculator$get_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
"similarity_factor"
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorSimilarityCalculator-get_valid_methods"></a>}}
\if{latex}{\out{\hypertarget{method-FactorSimilarityCalculator-get_valid_methods}{}}}
\subsection{Method \code{get_valid_methods()}}{
Get the valid methods for factor similarity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorSimilarityCalculator$get_valid_methods()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector of valid methods
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorSimilarityCalculator-calculate_similarity"></a>}}
\if{latex}{\out{\hypertarget{method-FactorSimilarityCalculator-calculate_similarity}{}}}
\subsection{Method \code{calculate_similarity()}}{
Calculate similarity between two factor values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorSimilarityCalculator$calculate_similarity(f1, f2, method, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f1}}{First factor value}

\item{\code{f2}}{Second factor value}

\item{\code{method}}{Method for comparison}

\item{\code{...}}{Additional arguments (not used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Normalized similarity score between 0 and 1
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorSimilarityCalculator-calculate_scores"></a>}}
\if{latex}{\out{\hypertarget{method-FactorSimilarityCalculator-calculate_scores}{}}}
\subsection{Method \code{calculate_scores()}}{
Calculate similarity scores between two lists of factors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorSimilarityCalculator$calculate_scores(list1, list2, method, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{list1}}{First list of factors}

\item{\code{list2}}{Second list of factors}

\item{\code{method}}{Method for comparison}

\item{\code{...}}{Additional arguments passed to calculate_similarity()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named vector of similarity scores
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorSimilarityCalculator-validate_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-FactorSimilarityCalculator-validate_inputs}{}}}
\subsection{Method \code{validate_inputs()}}{
Validate factor input lists
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorSimilarityCalculator$validate_inputs(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Lists of factors to validate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if valid, error otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorSimilarityCalculator-compare"></a>}}
\if{latex}{\out{\hypertarget{method-FactorSimilarityCalculator-compare}{}}}
\subsection{Method \code{compare()}}{
Compare factor similarity across multiple lists
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorSimilarityCalculator$compare(..., method = "exact")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Lists of factors to compare}

\item{\code{...}}{Additional arguments (not used)}

\item{\code{method}}{Character vector of similarity methods}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A SimilarityBase object with factor components
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorSimilarityCalculator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FactorSimilarityCalculator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorSimilarityCalculator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
