% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-formatters.R
\name{FactorSimilarityFormatter}
\alias{FactorSimilarityFormatter}
\title{Factor Similarity Formatter}
\description{
Formatter for factor similarity output
}
\section{Super class}{
\code{\link[samesies:SimilarityFormatterBase]{samesies::SimilarityFormatterBase}} -> \code{FactorSimilarityFormatter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FactorSimilarityFormatter-get_type}{\code{FactorSimilarityFormatter$get_type()}}
\item \href{#method-FactorSimilarityFormatter-print}{\code{FactorSimilarityFormatter$print()}}
\item \href{#method-FactorSimilarityFormatter-clone}{\code{FactorSimilarityFormatter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorSimilarityFormatter-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-FactorSimilarityFormatter-get_type}{}}}
\subsection{Method \code{get_type()}}{
Get the type of formatter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorSimilarityFormatter$get_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
"similarity_factor"
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorSimilarityFormatter-print"></a>}}
\if{latex}{\out{\hypertarget{method-FactorSimilarityFormatter-print}{}}}
\subsection{Method \code{print()}}{
Print a factor similarity object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorSimilarityFormatter$print(obj, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The similarity object to print}

\item{\code{...}}{Additional arguments (not used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the input object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorSimilarityFormatter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FactorSimilarityFormatter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorSimilarityFormatter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
