% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-plotters.R
\name{FactorSimilarityPlotter}
\alias{FactorSimilarityPlotter}
\title{Factor Similarity Plotter}
\description{
Plotter for factor similarity visualizations
}
\section{Super class}{
\code{\link[samesies:SimilarityPlotterBase]{samesies::SimilarityPlotterBase}} -> \code{FactorSimilarityPlotter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FactorSimilarityPlotter-get_type}{\code{FactorSimilarityPlotter$get_type()}}
\item \href{#method-FactorSimilarityPlotter-plot}{\code{FactorSimilarityPlotter$plot()}}
\item \href{#method-FactorSimilarityPlotter-clone}{\code{FactorSimilarityPlotter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorSimilarityPlotter-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-FactorSimilarityPlotter-get_type}{}}}
\subsection{Method \code{get_type()}}{
Get the type of plotter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorSimilarityPlotter$get_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
"similarity_factor"
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorSimilarityPlotter-plot"></a>}}
\if{latex}{\out{\hypertarget{method-FactorSimilarityPlotter-plot}{}}}
\subsection{Method \code{plot()}}{
Create a plot of factor similarity scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorSimilarityPlotter$plot(obj, type = "combined", palette = "Set2", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The similarity object to plot}

\item{\code{type}}{Plot type: "combined", "boxplot", "point", "violin", or "heatmap"}

\item{\code{palette}}{Color palette to use}

\item{\code{...}}{Additional arguments (not used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorSimilarityPlotter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FactorSimilarityPlotter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorSimilarityPlotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
