% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-formatters.R
\name{NumberSimilarityFormatter}
\alias{NumberSimilarityFormatter}
\title{Number Similarity Formatter}
\description{
Formatter for number similarity output
}
\section{Super class}{
\code{\link[samesies:SimilarityFormatterBase]{samesies::SimilarityFormatterBase}} -> \code{NumberSimilarityFormatter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NumberSimilarityFormatter-get_type}{\code{NumberSimilarityFormatter$get_type()}}
\item \href{#method-NumberSimilarityFormatter-print}{\code{NumberSimilarityFormatter$print()}}
\item \href{#method-NumberSimilarityFormatter-clone}{\code{NumberSimilarityFormatter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumberSimilarityFormatter-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-NumberSimilarityFormatter-get_type}{}}}
\subsection{Method \code{get_type()}}{
Get the type of formatter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumberSimilarityFormatter$get_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
"similarity_number"
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumberSimilarityFormatter-print"></a>}}
\if{latex}{\out{\hypertarget{method-NumberSimilarityFormatter-print}{}}}
\subsection{Method \code{print()}}{
Print a number similarity object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumberSimilarityFormatter$print(obj, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The similarity object to print}

\item{\code{...}}{Additional arguments (not used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the input object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumberSimilarityFormatter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NumberSimilarityFormatter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumberSimilarityFormatter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
