% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-plotters.R
\name{NumberSimilarityPlotter}
\alias{NumberSimilarityPlotter}
\title{Number Similarity Plotter}
\description{
Plotter for number similarity visualizations
}
\section{Super class}{
\code{\link[samesies:SimilarityPlotterBase]{samesies::SimilarityPlotterBase}} -> \code{NumberSimilarityPlotter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NumberSimilarityPlotter-get_type}{\code{NumberSimilarityPlotter$get_type()}}
\item \href{#method-NumberSimilarityPlotter-plot}{\code{NumberSimilarityPlotter$plot()}}
\item \href{#method-NumberSimilarityPlotter-clone}{\code{NumberSimilarityPlotter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumberSimilarityPlotter-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-NumberSimilarityPlotter-get_type}{}}}
\subsection{Method \code{get_type()}}{
Get the type of plotter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumberSimilarityPlotter$get_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
"similarity_number"
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumberSimilarityPlotter-plot"></a>}}
\if{latex}{\out{\hypertarget{method-NumberSimilarityPlotter-plot}{}}}
\subsection{Method \code{plot()}}{
Create a plot of number similarity scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumberSimilarityPlotter$plot(obj, type = "combined", palette = "Set2", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The similarity object to plot}

\item{\code{type}}{Plot type: "combined", "boxplot", "point", "violin", or "scatter"}

\item{\code{palette}}{Color palette to use}

\item{\code{...}}{Additional arguments (not used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumberSimilarityPlotter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NumberSimilarityPlotter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumberSimilarityPlotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
