% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar.R
\name{pair_averages.similar}
\alias{pair_averages.similar}
\title{Calculate average similarity by pair}
\usage{
\method{pair_averages}{similar}(x, method = NULL, ...)
}
\arguments{
\item{x}{A similar object}

\item{method}{Optional character vector of methods to include}
}
\value{
Data frame of pair-wise averages
}
\description{
Calculate average similarity by pair
}
