% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-base.R
\name{pair_averages.SimilarityBase}
\alias{pair_averages.SimilarityBase}
\title{Method implementation for pair_averages to work with SimilarityBase}
\usage{
\method{pair_averages}{SimilarityBase}(x, method = NULL, ...)
}
\description{
Method implementation for pair_averages to work with SimilarityBase
}
