% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same-factor.R
\name{plot.default}
\alias{plot.default}
\title{Default S3 method for plot}
\usage{
\method{plot}{default}(x, ...)
}
\arguments{
\item{x}{Object to plot}

\item{...}{Additional arguments passed to graphics::plot}
}
\value{
Result of graphics::plot
}
\description{
Default S3 method for plot
}
