% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same-factor.R
\name{plot}
\alias{plot}
\title{S3 generic for plot}
\usage{
plot(x, ...)
}
\arguments{
\item{x}{Object to plot}

\item{...}{Additional arguments passed to specific methods}
}
\value{
A plot object, typically a ggplot2 object for samesies classes
}
\description{
S3 generic for plot
}
