% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-base.R
\name{plot.SimilarityBase}
\alias{plot.SimilarityBase}
\title{Method implementation for plot to work with SimilarityBase}
\usage{
\method{plot}{SimilarityBase}(x, type = "combined", palette = "Set2", ...)
}
\description{
Method implementation for plot to work with SimilarityBase
}
