% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same-class.R
\name{plot_methods}
\alias{plot_methods}
\alias{plot_methods.similar}
\title{Generic method for plot methods}
\usage{
plot_methods(x, type = "default", ...)

\method{plot_methods}{similar}(x, type = "combined", palette = "Set2", ...)
}
\arguments{
\item{x}{A similarity object}

\item{type}{The type of plot to create}

\item{...}{Additional arguments passed to specific methods}
}
\value{
A ggplot2 object
}
\description{
Generic method for plot methods
}
