% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same-class.R
\name{plot_methods.similar_factor}
\alias{plot_methods.similar_factor}
\title{Plot method for similar_factor objects}
\usage{
\method{plot_methods}{similar_factor}(x, type = "default", palette = "Set2", ...)
}
\arguments{
\item{x}{A similar_factor object}

\item{type}{Plot type to create: "default", "heatmap", "boxplot", or "point"}

\item{palette}{Color palette to use for plot elements}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
A ggplot2 object visualizing factor similarity scores
}
\description{
Plot method for similar_factor objects
}
