% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same-class.R
\name{plot_methods.similar_number}
\alias{plot_methods.similar_number}
\title{Plot method for similar_number objects}
\usage{
\method{plot_methods}{similar_number}(x, type = "default", palette = "Set2", ...)
}
\arguments{
\item{x}{A similar_number object}

\item{type}{Plot type to create: "default", "histogram", "boxplot", or "point"}

\item{palette}{Color palette to use for plot elements}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
A ggplot2 object visualizing numeric similarity scores
}
\description{
Plot method for similar_number objects
}
