% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same-number.R
\name{print.same_number}
\alias{print.same_number}
\title{Print Method for same_number Objects}
\usage{
\method{print}{same_number}(x, ...)
}
\arguments{
\item{x}{A same_number object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for same_number Objects
}
