% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-s3-bridge.R
\name{print.similar}
\alias{print.similar}
\title{Print method for similar objects}
\usage{
\method{print}{similar}(x, ...)
}
\arguments{
\item{x}{A similar object}

\item{...}{Additional arguments (not used)}
}
\value{
The object invisibly
}
\description{
Print method for similar objects
}
