% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-similar-factor-methods.R
\name{print.similar_factor}
\alias{print.similar_factor}
\title{Print method for similar_factor objects}
\usage{
\method{print}{similar_factor}(x, ...)
}
\arguments{
\item{x}{A similar_factor object}

\item{...}{Additional arguments (not used)}
}
\value{
The object invisibly
}
\description{
Print method for similar_factor objects
}
