% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-similar-number-methods.R
\name{print.similar_number}
\alias{print.similar_number}
\title{Print method for similar_number objects}
\usage{
\method{print}{similar_number}(x, ...)
}
\arguments{
\item{x}{A similar_number object}

\item{...}{Additional arguments (not used)}
}
\value{
The object invisibly
}
\description{
Print method for similar_number objects
}
