% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-similar-base-methods.R
\name{print.summary.similar}
\alias{print.summary.similar}
\title{Print method for summary.similar objects}
\usage{
\method{print}{summary.similar}(x, ...)
}
\arguments{
\item{x}{A summary.similar object}

\item{...}{Additional arguments (not used)}
}
\value{
The summary object invisibly
}
\description{
Print method for summary.similar objects
}
