% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-similar-factor-methods.R
\name{print.summary.similar_factor}
\alias{print.summary.similar_factor}
\title{Print method for summary.similar_factor objects}
\usage{
\method{print}{summary.similar_factor}(x, ...)
}
\arguments{
\item{x}{A summary.similar_factor object}

\item{...}{Additional arguments (not used)}
}
\value{
The object invisibly
}
\description{
Print method for summary.similar_factor objects
}
