% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same-text.R
\name{print.summary.similar_text}
\alias{print.summary.similar_text}
\title{Print Method for summary.similar_text Objects}
\usage{
\method{print}{summary.similar_text}(x, ...)
}
\arguments{
\item{x}{A summary.similar_text object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for summary.similar_text Objects
}
