% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-calculators.R
\name{SimilarityCalculatorBase}
\alias{SimilarityCalculatorBase}
\title{Base Similarity Calculator}
\description{
\code{SimilarityCalculatorBase} is an R6 class that defines the interface for
similarity calculators.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SimilarityCalculatorBase-get_type}{\code{SimilarityCalculatorBase$get_type()}}
\item \href{#method-SimilarityCalculatorBase-get_valid_methods}{\code{SimilarityCalculatorBase$get_valid_methods()}}
\item \href{#method-SimilarityCalculatorBase-validate}{\code{SimilarityCalculatorBase$validate()}}
\item \href{#method-SimilarityCalculatorBase-calculate_similarity}{\code{SimilarityCalculatorBase$calculate_similarity()}}
\item \href{#method-SimilarityCalculatorBase-calculate_scores}{\code{SimilarityCalculatorBase$calculate_scores()}}
\item \href{#method-SimilarityCalculatorBase-validate_inputs}{\code{SimilarityCalculatorBase$validate_inputs()}}
\item \href{#method-SimilarityCalculatorBase-compare}{\code{SimilarityCalculatorBase$compare()}}
\item \href{#method-SimilarityCalculatorBase-clone}{\code{SimilarityCalculatorBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimilarityCalculatorBase-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-SimilarityCalculatorBase-get_type}{}}}
\subsection{Method \code{get_type()}}{
Get the type of calculator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimilarityCalculatorBase$get_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string identifying the calculator type
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimilarityCalculatorBase-get_valid_methods"></a>}}
\if{latex}{\out{\hypertarget{method-SimilarityCalculatorBase-get_valid_methods}{}}}
\subsection{Method \code{get_valid_methods()}}{
Get the valid methods for this calculator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimilarityCalculatorBase$get_valid_methods()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector of valid methods
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimilarityCalculatorBase-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SimilarityCalculatorBase-validate}{}}}
\subsection{Method \code{validate()}}{
Validate that an object has valid methods for this calculator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimilarityCalculatorBase$validate(obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The similarity object to validate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL if valid, error message otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimilarityCalculatorBase-calculate_similarity"></a>}}
\if{latex}{\out{\hypertarget{method-SimilarityCalculatorBase-calculate_similarity}{}}}
\subsection{Method \code{calculate_similarity()}}{
Calculate similarity between two elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimilarityCalculatorBase$calculate_similarity(elem1, elem2, method, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{elem1}}{First element}

\item{\code{elem2}}{Second element}

\item{\code{method}}{The method to use}

\item{\code{...}}{Additional arguments passed to method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A similarity score between 0 and 1
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimilarityCalculatorBase-calculate_scores"></a>}}
\if{latex}{\out{\hypertarget{method-SimilarityCalculatorBase-calculate_scores}{}}}
\subsection{Method \code{calculate_scores()}}{
Calculate similarity scores between two lists
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimilarityCalculatorBase$calculate_scores(list1, list2, method, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{list1}}{First list}

\item{\code{list2}}{Second list}

\item{\code{method}}{The method to use}

\item{\code{...}}{Additional arguments passed to method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named vector of similarity scores
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimilarityCalculatorBase-validate_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-SimilarityCalculatorBase-validate_inputs}{}}}
\subsection{Method \code{validate_inputs()}}{
Validate inputs before calculation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimilarityCalculatorBase$validate_inputs(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Lists to validate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if valid, error otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimilarityCalculatorBase-compare"></a>}}
\if{latex}{\out{\hypertarget{method-SimilarityCalculatorBase-compare}{}}}
\subsection{Method \code{compare()}}{
Compare multiple lists and return a similarity object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimilarityCalculatorBase$compare(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Lists to compare}

\item{\code{...}}{Additional arguments passed to method}

\item{\code{method}}{Methods to use for comparison}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A similarity object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimilarityCalculatorBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SimilarityCalculatorBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimilarityCalculatorBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
