% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-formatters.R
\name{SimilarityFormatterBase}
\alias{SimilarityFormatterBase}
\title{Base Similarity Formatter}
\description{
\code{SimilarityFormatterBase} is an R6 class that defines the interface for
similarity formatters that handle output display.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SimilarityFormatterBase-get_type}{\code{SimilarityFormatterBase$get_type()}}
\item \href{#method-SimilarityFormatterBase-print}{\code{SimilarityFormatterBase$print()}}
\item \href{#method-SimilarityFormatterBase-clone}{\code{SimilarityFormatterBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimilarityFormatterBase-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-SimilarityFormatterBase-get_type}{}}}
\subsection{Method \code{get_type()}}{
Get the type of formatter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimilarityFormatterBase$get_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string identifying the formatter type
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimilarityFormatterBase-print"></a>}}
\if{latex}{\out{\hypertarget{method-SimilarityFormatterBase-print}{}}}
\subsection{Method \code{print()}}{
Print a similarity object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimilarityFormatterBase$print(obj, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The similarity object to print}

\item{\code{...}}{Additional arguments (not used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the input object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimilarityFormatterBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SimilarityFormatterBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimilarityFormatterBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
