% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-plotters.R
\name{SimilarityPlotterBase}
\alias{SimilarityPlotterBase}
\title{Base Similarity Plotter}
\description{
\code{SimilarityPlotterBase} is an R6 class that defines the interface for
similarity plotters that create visualizations.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SimilarityPlotterBase-get_type}{\code{SimilarityPlotterBase$get_type()}}
\item \href{#method-SimilarityPlotterBase-plot}{\code{SimilarityPlotterBase$plot()}}
\item \href{#method-SimilarityPlotterBase-clone}{\code{SimilarityPlotterBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimilarityPlotterBase-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-SimilarityPlotterBase-get_type}{}}}
\subsection{Method \code{get_type()}}{
Get the type of plotter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimilarityPlotterBase$get_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string identifying the plotter type
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimilarityPlotterBase-plot"></a>}}
\if{latex}{\out{\hypertarget{method-SimilarityPlotterBase-plot}{}}}
\subsection{Method \code{plot()}}{
Create a plot of similarity scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimilarityPlotterBase$plot(obj, type = "combined", palette = "Set2", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The similarity object to plot}

\item{\code{type}}{Plot type, varies by plotter}

\item{\code{palette}}{Color palette to use}

\item{\code{...}}{Additional arguments passed to specific plotters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimilarityPlotterBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SimilarityPlotterBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimilarityPlotterBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
