% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same-factor.R
\name{summary.same_factor}
\alias{summary.same_factor}
\title{Summary Method for same_factor Objects}
\usage{
\method{summary}{same_factor}(object, ...)
}
\arguments{
\item{object}{A same_factor object}

\item{...}{Additional arguments (not used)}
}
\value{
A summary.same_factor object
}
\description{
Summary Method for same_factor Objects
}
