% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-similar-base-methods.R
\name{summary.similar}
\alias{summary.similar}
\title{Summarize a similarity object}
\usage{
\method{summary}{similar}(object, ...)
}
\arguments{
\item{object}{A similarity object}

\item{...}{Additional arguments (not used)}
}
\value{
A summary object
}
\description{
Summarize a similarity object
}
