% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-similar-factor-methods.R
\name{summary.similar_factor}
\alias{summary.similar_factor}
\title{Summary method for similar_factor objects}
\usage{
\method{summary}{similar_factor}(object, ...)
}
\arguments{
\item{object}{A similar_factor object}

\item{...}{Additional arguments (not used)}
}
\value{
A summary.similar_factor object
}
\description{
Summary method for similar_factor objects
}
