% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-similar-number-methods.R
\name{summary.similar_number}
\alias{summary.similar_number}
\title{Summary method for similar_number objects}
\usage{
\method{summary}{similar_number}(object, ...)
}
\arguments{
\item{object}{A similar_number object}

\item{...}{Additional arguments (not used)}
}
\value{
A summary.similar_number object
}
\description{
Summary method for similar_number objects
}
