% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-similar-text-methods.R
\name{summary.similar_text}
\alias{summary.similar_text}
\title{Summary method for similar_text objects}
\usage{
\method{summary}{similar_text}(object, ...)
}
\arguments{
\item{object}{A similar_text object}

\item{...}{Additional arguments (not used)}
}
\value{
A summary.similar_text object
}
\description{
Summary method for similar_text objects
}
