% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-base.R
\name{summary.SimilarityBase}
\alias{summary.SimilarityBase}
\title{Method implementation for summary to work with SimilarityBase}
\usage{
\method{summary}{SimilarityBase}(object, ...)
}
\description{
Method implementation for summary to work with SimilarityBase
}
