% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-formatters.R
\name{TextSimilarityFormatter}
\alias{TextSimilarityFormatter}
\title{Text Similarity Formatter}
\description{
Formatter for text similarity output
}
\section{Super class}{
\code{\link[samesies:SimilarityFormatterBase]{samesies::SimilarityFormatterBase}} -> \code{TextSimilarityFormatter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TextSimilarityFormatter-get_type}{\code{TextSimilarityFormatter$get_type()}}
\item \href{#method-TextSimilarityFormatter-print}{\code{TextSimilarityFormatter$print()}}
\item \href{#method-TextSimilarityFormatter-clone}{\code{TextSimilarityFormatter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextSimilarityFormatter-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-TextSimilarityFormatter-get_type}{}}}
\subsection{Method \code{get_type()}}{
Get the type of formatter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextSimilarityFormatter$get_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
"similarity_text"
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextSimilarityFormatter-print"></a>}}
\if{latex}{\out{\hypertarget{method-TextSimilarityFormatter-print}{}}}
\subsection{Method \code{print()}}{
Print a text similarity object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextSimilarityFormatter$print(obj, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The similarity object to print}

\item{\code{...}}{Additional arguments (not used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the input object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextSimilarityFormatter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TextSimilarityFormatter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextSimilarityFormatter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
