% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-plotters.R
\name{TextSimilarityPlotter}
\alias{TextSimilarityPlotter}
\title{Text Similarity Plotter}
\description{
Plotter for text similarity visualizations
}
\section{Super class}{
\code{\link[samesies:SimilarityPlotterBase]{samesies::SimilarityPlotterBase}} -> \code{TextSimilarityPlotter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TextSimilarityPlotter-get_type}{\code{TextSimilarityPlotter$get_type()}}
\item \href{#method-TextSimilarityPlotter-plot}{\code{TextSimilarityPlotter$plot()}}
\item \href{#method-TextSimilarityPlotter-clone}{\code{TextSimilarityPlotter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextSimilarityPlotter-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-TextSimilarityPlotter-get_type}{}}}
\subsection{Method \code{get_type()}}{
Get the type of plotter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextSimilarityPlotter$get_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
"similarity_text"
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextSimilarityPlotter-plot"></a>}}
\if{latex}{\out{\hypertarget{method-TextSimilarityPlotter-plot}{}}}
\subsection{Method \code{plot()}}{
Create a plot of text similarity scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextSimilarityPlotter$plot(obj, type = "combined", palette = "Set2", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The similarity object to plot}

\item{\code{type}}{Plot type: "combined", "boxplot", "point", or "violin"}

\item{\code{palette}}{Color palette to use}

\item{\code{...}}{Additional arguments (not used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextSimilarityPlotter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TextSimilarityPlotter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextSimilarityPlotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
