% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{validate_number_inputs}
\alias{validate_number_inputs}
\title{Validate Numeric List Inputs}
\usage{
validate_number_inputs(...)
}
\arguments{
\item{...}{Lists to validate}
}
\value{
TRUE if validation passes, otherwise throws an error
}
\description{
Validates that all inputs are lists containing only numeric values
}
\keyword{internal}
