% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-same-number.R
\name{auto_epsilon}
\alias{auto_epsilon}
\title{Calculate Epsilon Value for Fuzzy Matching}
\usage{
auto_epsilon(values, percentile = 0.1)
}
\arguments{
\item{values}{Numeric vector or list of values}

\item{percentile}{Percentile used for epsilon calculation (default: 0.1)}
}
\value{
Numeric epsilon value between min_epsilon and max_epsilon
}
\description{
Calculate Epsilon Value for Fuzzy Matching
}
\keyword{internal}
