% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000.R, R/same-class.R
\name{average_similarity}
\alias{average_similarity}
\alias{average_similarity.similar}
\title{Calculate Average Similarity Scores}
\usage{
average_similarity(x, ...)

\method{average_similarity}{similar}(x, ...)
}
\arguments{
\item{x}{A similarity object created by one of the similarity functions}

\item{...}{Additional arguments passed to specific methods}
}
\value{
A named numeric vector of mean similarity scores for each method
}
\description{
Calculates and returns the average similarity score for each method used in the comparison.
}
