% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar.R
\name{average_similarity.similar}
\alias{average_similarity.similar}
\title{Calculate average similarity across all methods}
\usage{
\method{average_similarity}{similar}(x, ...)
}
\arguments{
\item{x}{A similar object}

\item{...}{Additional arguments (not used)}
}
\value{
Named numeric vector of means by method
}
\description{
Calculate average similarity across all methods
}
