% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R, R/2-similar-base-methods.R
\name{pair_averages}
\alias{pair_averages}
\title{Calculate Average Similarity Scores By Pairs}
\usage{
pair_averages(x, method = NULL, ...)

pair_averages(x, method = NULL, ...)
}
\arguments{
\item{x}{A similarity object}

\item{method}{Optional character vector of methods to include}

\item{...}{Additional arguments (not used)}
}
\value{
A data frame containing:
\item{method}{The similarity method used}
\item{pair}{The pair of lists compared}
\item{avg_score}{Mean similarity score for the pair}

A data frame containing pair-wise average scores
}
\description{
Calculates and returns the average similarity scores for each pair of lists
compared, broken down by method.
}
