% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000.R, R/same-class.R
\name{pair_averages}
\alias{pair_averages}
\alias{pair_averages.similar}
\title{Calculate Average Similarity Scores By Pairs}
\usage{
pair_averages(x, method = NULL, ...)

\method{pair_averages}{similar}(x, method = NULL, ...)
}
\arguments{
\item{x}{A similarity object created by one of the similarity functions}

\item{method}{Optional character vector of methods to include. If NULL, uses all methods.}

\item{...}{Additional arguments passed to specific methods}
}
\value{
A data frame containing:
\item{method}{The similarity method used}
\item{pair}{The pair of lists compared}
\item{avg_score}{Mean similarity score for the pair}
}
\description{
Calculates and returns the average similarity scores for each pair of lists
compared, broken down by method.
}
