% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-s3-bridge.R
\name{pair_averages.similar}
\alias{pair_averages.similar}
\title{Calculate average similarity scores by pairs}
\usage{
\method{pair_averages}{similar}(x, method = NULL, ...)
}
\arguments{
\item{x}{A similarity object}

\item{method}{Optional character vector of methods to include}

\item{...}{Additional arguments (not used)}
}
\value{
A data frame containing pair-wise average scores
}
\description{
Calculate average similarity scores by pairs
}
