% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-similar-base-methods.R
\name{print.similar}
\alias{print.similar}
\title{Print a similarity object}
\usage{
\method{print}{similar}(x, ...)
}
\arguments{
\item{x}{A similarity object}

\item{...}{Additional arguments (not used)}
}
\value{
The object invisibly
}
\description{
Print a similarity object
}
