% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same-factor.R
\name{print.summary.similar_factor}
\alias{print.summary.similar_factor}
\title{Print Method for summary.similar_factor Objects}
\usage{
\method{print}{summary.similar_factor}(x, ...)
}
\arguments{
\item{x}{A summary.similar_factor object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for summary.similar_factor Objects
}
