% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-similar-text-methods.R
\name{print.summary.similar_text}
\alias{print.summary.similar_text}
\title{Print method for summary.similar_text objects}
\usage{
\method{print}{summary.similar_text}(x, ...)
}
\arguments{
\item{x}{A summary.similar_text object}

\item{...}{Additional arguments (not used)}
}
\value{
The object invisibly
}
\description{
Print method for summary.similar_text objects
}
