% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-parent-children.R
\name{similar_factor}
\alias{similar_factor}
\title{Factor similarity comparison class}
\usage{
similar_factor(
  scores = list(),
  summary = list(),
  methods = character(0),
  list_names = character(0),
  levels = character(0)
)
}
\description{
\code{similar_factor} is an S7 class for categorical/factor similarity comparisons.
}
\details{
This class extends the abstract \code{similar} class and implements
categorical data-specific similarity comparison methods.
}
