% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-parent-children.R
\name{similar_number}
\alias{similar_number}
\title{Numeric similarity comparison class}
\usage{
similar_number(
  scores = list(),
  summary = list(),
  methods = character(0),
  list_names = character(0),
  raw_values = list()
)
}
\description{
\code{similar_number} is an S7 class for numeric similarity comparisons.
}
\details{
This class extends the abstract \code{similar} class and implements
numeric data-specific similarity comparison methods.
}
