% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-parent-children.R
\name{similar_text}
\alias{similar_text}
\title{Text similarity comparison class}
\usage{
similar_text(
  scores = list(),
  summary = list(),
  methods = character(0),
  list_names = character(0)
)
}
\description{
\code{similar_text} is an S7 class for text similarity comparisons.
}
\details{
This class extends the abstract \code{similar} class and implements
text-specific similarity comparison methods.
}
